address consistency checks
