+++ workflow/Snakefile
@@ -83,7 +83,6 @@ wildcard_constraints:
 
 
 # Merge subworkflow configs and main config
-configfile: "config/config.county.yaml"
 configfile: "config/config.cluster.yaml"
 configfile: "config/config.common.yaml"
 configfile: "config/config.plotting.yaml"
